extract instances:

+++ src/mscoco_extract_instances.py
@@ -92,7 +92,7 @@ def polyFill(img, width, height, segmentation):
         print 'SegX:', seg[0:N:2]
         poly = np.concatenate(
             (np.array(seg[1:N:2]).reshape(N/2, 1), 
-             np.array(seg[0:N:2]).reshape(N/2, 1)), axis=-1)
+             np.array(seg[0:N:2]).reshape(N/2, 1)), axis=-1).astype(int)
         print 'Poly', poly, poly.shape
         polys.append(poly)
     cv2.fillPoly(img=img, pts=np.array(polys), color=(1, 1, 1))
